FARRAGO: "N-way join" planner refactoring.
Several relexps have system columns. (Also validators and type factories.) The
list of system columns is for sub-projects; it is generally empty (not null) in
luciddb. System fields, if present, are always on the leading edge of a
relexp's rowType. Joins are more complex: there is the new system field, then
the left input's system field, on the leading edge of the output rowType, but
the system fields of the right input will appear in the middle of the output
rowType.

During validation, all possible system fields are present; if the fields are
not applicable for a particular relexp, the validator will find the field, but
give an error. Later in planning, the fields may be stripped away, and just
those system fields that are used are retained. (See trimming below.) Change
several rules to accommodate system fields; typically the accommodation is to
add or subtract an offset.

After sql-to-rel translation, and before planning, trim unused fields
throughout the relexp tree. This yields more efficient plans, removes system
fields that are not applicable for a particular relexp type, and also speeds
the planning process. It cannot be done as a planner rule, because it changes
the output row type.

In aggregate relexps, grouping columns are now represented as a bitset, not a
count. Therefore they do not need to be on the leading edge. System columns are
not regarded as grouping columns, even though their values they may pass
through the relexp unchanged.

For some enum types, use '==' or a switch statement rather than .equals().

Add methods Util.bitSetOf, .bitSetBetween, .toArray(BitSet), .toIter(BitSet)
for more easily handling bit sets.

Add method "RelNode.copy(RelNode... children)" that unlike clone(), allows
child relexps to be replaced. Some relexps also have variants with more
parameters.

Add RelOptRuleOperand.Predicate, that allows operands to match based on
arbitrary side-conditions such as trait of the matched node, or whether it has
system fields. It is more efficient than applying the side-conditions in the
rule's onMatch method.

A few planner methods, such as onRegister, now take a RelOptRuleCall as a
parameter. This allows construction of a planning history; each RelNode knows
which rule call created it, hence its ancestors. If planning fails, you can
dump the planning history from in-memory state. You can get away with MUCH less
verbose tracing.

Migrate a few data structures from arrays to lists, because lists (being
interfaces) allow a functional programming style. In a similar vein, change
TypeFactory.createStructType(RelDataType[], String[]) to
.createStructType(List<Map.Entry<String, RelDataType>>). Note that
RelDataTypeField implements Map.Entry<String, RelDataType>, so lists of fields
can be passed to such methods. Use of interface FieldInfo is now discouraged,
because a list of fields does just as well.

Add FennelRelUtil.isFennelBuiltinAggFunction, to test whether an agg function
is built into fennel.

Move InputFinder.apply to RexUtil, so it can be used for any RexVisitor.

Fix saffron for JDK 1.6 (specifically JDBC 4.0).

Clean up how OJPreparingStmt is subclassed; previously its implement method was
private, therefore derived classes could not override it. The new implement0
method allows subclasses to control which parameters get passed to implement.

[git-p4: depot-paths = "//open/dt/dev/": change = 14416]